Add ``AnimateDrawCommand``
This is meant to act as some kind of "simplification" of the rendering backend of flixel-animate. Up until now, i kept adding more and more args on top of eachother to render different elements, this sucked and was hard to work with on the long run. Now it should be much cleaner and easy to work upon.

This also brings some extra functionality to more classes! Now atlas instances should also have blend modes and color transforms available to edit. Might not be accurate to flash, but its useful.

Im making a new branch because this is still highly experimental, both in code practice and in performance, so it may not make the final cut.